--- packages/marketing-cloud-logo/src/utils/index.ts
@@ -1,2 +1,0 @@
-export {calcSvgPadding} from './calcSvgPadding';
-export {handleColor} from './handleColor';
