+++ packages/marketing-cloud-logo/tsconfig.json
@@ -22,6 +22,6 @@
     "target": "ES2017",
     "typeRoots": ["./src/types", "./node_modules/@types"]
   },
-  "include": ["./src/**/*"],
+  "include": ["./src"],
   "exclude": ["node_modules"]
 }
